Update custom_field.php
Noticed (through the API) that if you try to update 'params' you receive an error: "must be an object." getDaoFieldsFromKeyAndValue expects an object it then encodes into JSON, but if you hand it already-encoded json, it doesn't recognize it.  This would both recognize and validate a json encoded string as input.