* [Cards/Widgets] The ‘Record Fields’ card widget displays built-in fields, custom fields, and search buttons for any record.
The displayed record can be specified using placeholders. For instance, on a ticket profile, the fields of the ticket and its organization can be displayed at the same time.

The fields and fieldsets to display are configurable. Fields can be dragged into a desired order. This drastically cleans up profiles by only displaying fields that are important and hiding everything else.

A ‘Show record links’ option determines whether or not to show the ‘Links’ section for adding links and showing a search button for each type of linked record.

A ‘Show empty fields’ option determines whether blank fields will be included in the output. This is disabled by default (i.e. fields with no value are hidden).

Custom search counter buttons can be configured for each instance of the widget. These buttons run a search query using placeholders from the target record and return the number of results. For instance, a different set of search counters could be displayed on an organization fields widget on tickets and on contacts. For tickets, counts of the organization’s ticket history and contacts could be shown.