[Platform/Security] Refactored the anti-CSRF protections.
GET requests may no longer include the CSRF token in the URL, as it could (theoretically) leak in the URL. Only the HTTP headers or POST fields are checked for the token. Previously, the origin/referer of a request was only checked when it included a `c` (controller) or `a` argument. Now the origin of all non-GET URLs are checked regardless, as well as all Ajax requests, or any requests including controller/action routing. Previously the CSRF token was only compared on requests including a controller or action. Now the CSRF token is checked on all non-GET requests (excluding `/cron`, `/oauth`, `/portal`, `/sso`, and `/webhooks`).