10.4.2: * [Roles] Fixed a permissions check bug when a non-admin worker is a member of a role with 'All' permissions. Depending on the order the roles are checked, this can return `void` rather than the expected `true`. Thanks to [@tomryder-inspirenet](https://github.com/tomryder-inspirenet) for the bug report and patch!
Fixes #1739

When `DAO_Worker::hasPriv()` is called with
e.g. `'contexts.cerberusweb.contexts.ticket.create'` in
`PageSection_ProfilesDraft::_profileAction_saveComposePeek()`, and the
worker is not a super-user but is in a role with "All" permissions
(e.g. "Default"), the caller checks for the privilege key "*", which is
not set if there's a short-circuit void return like this.  It looks to
me like returning the `$privs` object as soon as the loop hits a role
with "all"/"everything" privileges was what was actually intended.